Move sheet to separate object.
