+++ opal/fron/core.rb
@@ -1,3 +1,4 @@
 require 'fron/core/logger'
 require 'fron/core/component'
 require 'fron/core/eventable'
+require 'fron/core/sheet'
